+++ src/app/[locale]/faq/page.tsx
@@ -1,4 +1,4 @@
-import { FaqSectionCta } from "@/src/components/faqPage/FaqSectionCta";
+import { FaqSectionCta } from "@/src/components/faqPage/сtaSection/FaqSectionCta";
 
 export default function FAQ() {
     return (
